project101repo/requirements.txt
# external requirements
awscli

click
coverage
flake8
matplotlib
mlflow
numpy
pandas
pyyaml
python-dotenv>=0.5.1
requests
scikit-learn
seaborn
Sphinx
streamlit
streamlit==1.24
streamlit-extras
#tensorflow
textblob
transformers
#tf-keras
torch
spacy
sentence_transformers

#RAG
langchain
openpyxl
langchain-core
langchain_community
langchain-huggingface
faiss-cpu
sentence-transformers